+++ carp_mobile_sensing/lib/sampling_packages/sensors/sensor_probes.dart
@@ -24,7 +24,7 @@ abstract class BufferingSensorProbe extends BufferingPeriodicStreamProbe {
   void onSamplingEnd() {}
 }
 
-/// A  probe collecting raw data from the accelerometer.
+/// A probe collecting raw data from the accelerometer.
 ///
 /// Note that this probe generates a lot of data and should be used
 /// with caution.
@@ -41,7 +41,7 @@ class AccelerometerProbe extends StreamProbe {
 /// setting the [interval] and [duration] of the sampling rate.
 class BufferingAccelerometerProbe extends BufferingSensorProbe {
   @override
-  Stream<dynamic> get bufferingStream => accelerometerEvents;
+  Stream<AccelerometerEvent> get bufferingStream => accelerometerEvents;
 
   @override
   void onSamplingData(event) {
@@ -52,7 +52,35 @@ class BufferingAccelerometerProbe extends BufferingSensorProbe {
   }
 }
 
-/// A  probe collecting raw data from the gyroscope.
+/// A probe collecting raw data from the accelerometer.
+///
+/// Note that this probe generates a lot of data and should be used
+/// with caution.
+class UserAccelerometerProbe extends StreamProbe {
+  @override
+  Stream<Datum> get stream => userAccelerometerEvents
+      .map((event) => UserAccelerometerDatum.fromUserAccelerometerEvent(event));
+}
+
+/// A probe that collects accelerometer events and buffers them and return
+/// a [MultiDatum] with all the buffered [AccelerometerDatum]s.
+///
+/// See [PeriodicSamplingConfiguration] on how to configure this probe, including
+/// setting the [interval] and [duration] of the sampling rate.
+class BufferingUserAccelerometerProbe extends BufferingSensorProbe {
+  @override
+  Stream<UserAccelerometerEvent> get bufferingStream => userAccelerometerEvents;
+
+  @override
+  void onSamplingData(event) {
+    if (event is UserAccelerometerEvent) {
+      datum.addDatum(
+          UserAccelerometerDatum.fromUserAccelerometerEvent(event, multiDatum: true));
+    }
+  }
+}
+
+/// A probe collecting raw data from the gyroscope.
 ///
 /// Note that this probe generates a lot of data and should be used
 /// with caution.
@@ -69,13 +97,41 @@ class GyroscopeProbe extends StreamProbe {
 /// setting the [interval] and [duration] of the sampling rate.
 class BufferingGyroscopeProbe extends BufferingSensorProbe {
   @override
-  Stream<dynamic> get bufferingStream => gyroscopeEvents;
+  Stream<GyroscopeEvent> get bufferingStream => gyroscopeEvents;
 
   @override
   void onSamplingData(dynamic event) {
     if (event is GyroscopeEvent) {
       datum
           .addDatum(GyroscopeDatum.fromGyroscopeEvent(event, multiDatum: true));
+    }
+  }
+}
+
+/// A probe collecting raw data from the gyroscope.
+///
+/// Note that this probe generates a lot of data and should be used
+/// with caution.
+class MagnetometerProbe extends StreamProbe {
+  @override
+  Stream<Datum> get stream => magnetometerEvents
+      .map((event) => MagnetometerDatum.fromMagnetometerEvent(event));
+}
+
+/// A probe that collects gyroscope events and buffers them and return
+/// a [MultiDatum] with all the buffered [GyroscopeDatum]s.
+///
+/// See [PeriodicSamplingConfiguration] on how to configure this probe, including
+/// setting the [interval] and [duration] of the sampling rate.
+class BufferingMagnetometerProbe extends BufferingSensorProbe {
+  @override
+  Stream<MagnetometerEvent> get bufferingStream => magnetometerEvents;
+
+  @override
+  void onSamplingData(dynamic event) {
+    if (event is MagnetometerEvent) {
+      datum.addDatum(
+          MagnetometerDatum.fromMagnetometerEvent(event, multiDatum: true));
     }
   }
 }