+++ packages/carp_apps_package/lib/app_probes.dart
@@ -43,20 +43,34 @@ class AppUsageProbe extends DatumProbe {
   @override
   Future<Datum> getDatum() async {
     // get the last mark - if null, go back as specified in history
-
     DateTime start = samplingConfiguration.lastTime ??
         DateTime.now().subtract(samplingConfiguration.past);
     DateTime end = DateTime.now();
 
     debug(
         'Collecting app usage - start: ${start.toUtc()}, end: ${end.toUtc()}');
-    List<AppUsageInfo> infos = await AppUsage.getAppUsage(start, end);
+    List<AppUsageInfo> infos = await AppUsage().getAppUsage(start, end);
 
+    // Create maps for several sets of information
     Map<String, int> usage = {};
+    Map<String, DateTime> startRange = {};
+    Map<String, DateTime> stopRange = {};
+    Map<String, DateTime> lastUseForeground = {};
+
+    // use full package name
     for (AppUsageInfo inf in infos) {
-      usage[inf.appName] = inf.usage.inSeconds;
+      usage[inf.packageName] = inf.usage.inSeconds;
+      startRange[inf.packageName] = inf.startDate;
+      stopRange[inf.packageName] = inf.endDate;
+      lastUseForeground[inf.packageName] = inf.lastForeground;
     }
 
-    return AppUsageDatum(start.toUtc(), end.toUtc())..usage = usage;
+    AppUsageDatum toReturn =
+        (AppUsageDatum(start.toUtc(), end.toUtc())..usage = usage);
+    toReturn.stopRange = stopRange;
+    toReturn.startRange = startRange;
+    toReturn.lastUseForeground = lastUseForeground;
+
+    return toReturn;
   }
 }