+++ packages/carp_survey_package/test/carp_survey_package_test.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
+import 'package:research_package/research_package.dart';
 import 'package:test/test.dart';
 
 import 'package:carp_serializable/carp_serializable.dart';
@@ -14,6 +15,7 @@ void main() {
   setUp(() {
     // Initialization of serialization
     CarpMobileSensing();
+    ResearchPackage();
 
     // register the survey sampling package
     SamplingPackageRegistry().register(SurveySamplingPackage());
@@ -45,7 +47,7 @@ void main() {
     protocol.addTriggeredTask(
         DelayedTrigger(delay: Duration(seconds: 30)),
         RPAppTask(
-            type: SurveyUserTask.WHO5_SURVEY_TYPE,
+            type: SurveyUserTask.SURVEY_TYPE,
             name: 'WHO-5 Survey',
             rpTask: who5Task)
           ..measures.add(Measure(type: DeviceSamplingPackage.DEVICE))
@@ -78,8 +80,7 @@ void main() {
 
     expect(protocol.ownerId, 'alex@uni.dk');
     expect(protocol.masterDevices.first.roleName, Smartphone.DEFAULT_ROLENAME);
-    expect((protocol.tasks.last as RPAppTask).type,
-        SurveyUserTask.WHO5_SURVEY_TYPE);
+    expect((protocol.tasks.last as RPAppTask).type, SurveyUserTask.SURVEY_TYPE);
     print(toJsonString(protocol));
   });
 }