+++ packages/carp_esense_package/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.40.1
+* update to `esense_flutter` v. 0.6.2
+
 ## 0.40.0
 
 * update to `carp_mobile_sensing` v. 0.40.0